moved other propety details repositories to the common folder
