+++ libs/common/src/index.ts
@@ -29,3 +29,11 @@ export * from './database/entities/property-type.entity';
 export * from './database/entities/property-category.entity';
 export * from './database/entities/property-status.entity';
 export * from './database/entities/property-purpose.entity';
+export * from './repositories/properties-category.repository';
+export * from './repositories/properties-purpose.repository';
+export * from './repositories/properties-status.repository';
+export * from './repositories/properties-type.repository';
+export * from './services/properties-category.service';
+export * from './services/properties-purpose.service';
+export * from './services/properties-status.service';
+export * from './services/properties-type.service';
