+++ libs/common/src/repositories/properties-category.repository.ts
