+++ libs/common/src/repositories/properties-purpose.repository.ts
