+++ libs/common/src/repositories/properties-status.repository.ts
