+++ libs/common/src/repositories/properties-type.repository.ts
