+++ libs/common/src/repositories/repositories.module.ts
@@ -5,6 +5,10 @@ import { FeaturesRepository } from './features.repository';
 import { OrganizationRolesRepository } from './organization-roles.repository';
 import { RolesRepository } from '../repositories/roles.repository';
 import { EntityManager } from 'typeorm';
+import { PropertyCategoryRepository } from './properties-category.repository';
+import { PropertyStatusRepository } from './properties-status.repository';
+import { PropertyTypeRepository } from './properties-type.repository';
+import { PropertyPurposeRepository } from './properties-purpose.repository';
 
 @Module({
 	providers: [
@@ -33,6 +37,26 @@ import { EntityManager } from 'typeorm';
 			useFactory: (em: EntityManager) => new OrganizationRolesRepository(em),
 			inject: [EntityManager],
 		},
+		{
+			provide: PropertyCategoryRepository,
+			useFactory: (em: EntityManager) => new PropertyCategoryRepository(em),
+			inject: [EntityManager],
+		},
+		{
+			provide: PropertyStatusRepository,
+			useFactory: (em: EntityManager) => new PropertyStatusRepository(em),
+			inject: [EntityManager],
+		},
+		{
+			provide: PropertyTypeRepository,
+			useFactory: (em: EntityManager) => new PropertyTypeRepository(em),
+			inject: [EntityManager],
+		},
+		{
+			provide: PropertyPurposeRepository,
+			useFactory: (em: EntityManager) => new PropertyPurposeRepository(em),
+			inject: [EntityManager],
+		},
 	],
 	exports: [
 		UserProfilesRepository,
@@ -40,6 +64,10 @@ import { EntityManager } from 'typeorm';
 		PermissionsRepository,
 		FeaturesRepository,
 		OrganizationRolesRepository,
+		PropertyCategoryRepository,
+		PropertyStatusRepository,
+		PropertyTypeRepository,
+		PropertyPurposeRepository,
 	],
 })
 export class RepositoriesModule {}
